Work on entering a group chat
