+++ ChatSecure/Classes/Controllers/OTREncryptionManager.m
@@ -241,7 +241,7 @@ NSString *const OTRMessageStateKey = @"OTREncryptionManagerMessageStateKey";
         buddy = [OTRBuddy fetchBuddyForUsername:username accountName:accountName transaction:transaction];
     }];
     
-    if(!buddy || buddy.status == OTRBuddyStatusOffline) {
+    if(!buddy || buddy.status == ThreadStatusOffline) {
         return NO;
     }
     else {
