+++ ChatSecure/Classes/Model/Yap Storage/OTRBuddyEnums.h
@@ -1,0 +1,16 @@
+//
+//  OTRBuddyEnums.h
+//  ChatSecure
+//
+//  Created by David Chiles on 10/19/15.
+//  Copyright © 2015 Chris Ballinger. All rights reserved.
+//
+
+typedef NS_ENUM(NSInteger, OTRBuddyStatus) {
+    OTRBuddyStatusOffline   = 4,
+    OTRBuddyStatusXa        = 3,
+    OTRBuddyStatusDnd       = 2,
+    OTRBuddyStatusAway      = 1,
+    OTRBuddyStatusAvailable = 0
+};
+
