+++ ChatSecure/Classes/Model/Yap Storage/OTRMessage.h
@@ -43,6 +43,8 @@ extern const struct OTRMessageEdges {
 
 - (BOOL)transportedSecurely;
 
+- (BOOL)messageRead;
+
 
 @end
