+++ ChatSecure/Classes/Model/Yap Storage/OTRXMPPRoom.swift
@@ -9,22 +9,36 @@
 import UIKit
 import YapDatabase
 
+@objc public enum ThreadStatus:Int {
+    case Available = 0
+    case Away = 1
+    case DoNotDisturb = 2
+    case ExtendedAway = 3
+    case Offline = 4
+}
+
 @objc public protocol OTRThreadOwner: NSObjectProtocol {
     
     func threadName() -> String
     func threadIdentifier() -> String
+    func threadCollection() -> String
     func threadAccountIdentifier() -> String
     func setCurrentMessageText(text:String?)
     func currentMessageText() -> String?
+    func lastMessageDate() -> NSDate?
+    func avatarImage() -> UIImage
+    func currentStatus() -> ThreadStatus
+    func lastMessageWithTransaction(transaction:YapDatabaseReadTransaction) -> OTRMesssageProtocol?
 }
 
-public class OTRXMPPRoom: OTRYapDatabaseObject, OTRThreadOwner {
+public class OTRXMPPRoom: OTRYapDatabaseObject {
     
     public var accountUniqueId:String?
     public var ownJID:String?
     public var jid:String?
     public var joined = false
     public var messageText:String?
+    public var lastRoomMessageDate:NSDate?
     override public var uniqueId:String {
         get {
             if let account = self.accountUniqueId {
@@ -36,12 +50,22 @@ public class OTRXMPPRoom: OTRYapDatabaseObject, OTRThreadOwner {
         }
     }
     
+    public class func createUniqueId(accountId:String, jid:String) -> String {
+        return accountId + jid
+    }
+}
+
+extension OTRXMPPRoom:OTRThreadOwner {
     public func threadName() -> String {
         return self.jid ?? ""
     }
     
     public func threadIdentifier() -> String {
         return self.uniqueId
+    }
+    
+    public func threadCollection() -> String {
+        return OTRXMPPRoom.collection()
     }
     
     public func threadAccountIdentifier() -> String {
@@ -56,123 +80,25 @@ public class OTRXMPPRoom: OTRYapDatabaseObject, OTRThreadOwner {
         return self.messageText
     }
     
-    public class func createUniqueId(accountId:String, jid:String) -> String {
-        return accountId + jid
+    public func lastMessageDate() -> NSDate? {
+        return self.lastRoomMessageDate
     }
-}
-
-public class OTRXMPPRoomOccupant: OTRYapDatabaseObject, YapDatabaseRelationshipNode {
     
-    public static let roomEdgeName = "OTRRoomOccupantEdgeName"
+    public func avatarImage() -> UIImage {
+        return OTRImages.avatarImageWithUniqueIdentifier(self.uniqueId, avatarData: nil, displayName: nil, username: self.jid)
+    }
     
-    public var available = false
-    
-    /** This is the JID of the participant as it's known in teh room  ie baseball_chat@conference.dukgo.com/user123 */
-    public var jid:String?
-    
-    /** This is the name your known as in the room. Seems to be username without domain */
-    public var roomName:String?
-    
-    /**When given by the server we get the room participants reall JID*/
-    public var realJID:String?
-    
-    public var roomUniqueId:String?
-    
-    public func yapDatabaseRelationshipEdges() -> [AnyObject]! {
-        if let roomID = self.roomUniqueId {
-            let relationship = YapDatabaseRelationshipEdge(name: OTRXMPPRoomOccupant.roomEdgeName, sourceKey: self.uniqueId, collection: OTRXMPPRoomOccupant.collection(), destinationKey: roomID, collection: OTRXMPPRoom.collection(), nodeDeleteRules: YDB_NodeDeleteRules.DeleteSourceIfDestinationDeleted)
-            return [relationship]
-        } else {
-            return []
+    public func currentStatus() -> ThreadStatus {
+        switch self.joined {
+        case true:
+            return .Available
+        default:
+            return .Offline
         }
     }
-}
-
-public class OTRXMPPRoomMessage: OTRYapDatabaseObject {
     
-    public static let roomEdgeName = "OTRRoomMesageEdgeName"
-    
-    public var roomJID:String?
-    
-    /** This is the full JID of the sender. This should be equal to the occupant.jid*/
-    public var senderJID:String?
-    public var displayName:String?
-    public var incoming = false
-    public var messageText:String?
-    public var messageDate:NSDate?
-    
-    public var roomUniqueId:String?
-    
-    public override var hash: Int {
-        get {
-           return super.hash
-        }
-    }
-}
-
-extension OTRXMPPRoomMessage:YapDatabaseRelationshipNode {
-    //MARK: YapRelationshipNode
-    public func yapDatabaseRelationshipEdges() -> [AnyObject]! {
-        
-        if let roomID = self.roomUniqueId {
-            let relationship = YapDatabaseRelationshipEdge(name: OTRXMPPRoomMessage.roomEdgeName, sourceKey: self.uniqueId, collection: OTRXMPPRoomMessage.collection(), destinationKey: roomID, collection: OTRXMPPRoom.collection(), nodeDeleteRules: YDB_NodeDeleteRules.DeleteSourceIfDestinationDeleted)
-            return [relationship]
-        } else {
-            return []
-        }
-    }
-}
-
-extension OTRXMPPRoomMessage:OTRMesssageProtocol {
-    
-    //MARK: OTRMessageProtocol
-    
-    public func threadId() -> String! {
-        return self.roomUniqueId
-    }
-    
-    public func messageIncoming() -> Bool {
-        return self.incoming
-    }
-    
-    public func messageMediaItemKey() -> String! {
+    public func lastMessageWithTransaction(transaction: YapDatabaseReadTransaction) -> OTRMesssageProtocol? {
+        //TODO
         return nil
     }
-    
-    public func messageError() -> NSError! {
-        return nil
-    }
-    
-    public func transportedSecurely() -> Bool {
-        return false;
-    }
-    
-    //MARK: JSQMessageData Protocol methods
-    
-    public func senderId() -> String! {
-        return self.senderJID
-    }
-    
-    public func senderDisplayName() -> String! {
-        return self.displayName
-    }
-    
-    public func date() -> NSDate! {
-        return self.messageDate
-    }
-    
-    public func isMediaMessage() -> Bool {
-        return false
-    }
-    
-    public func messageHash() -> UInt {
-        
-        //TODO this is not correct but UInt(self.hash) is not working
-        return UInt(self.date().timeIntervalSince1970)
-    }
-    
-    public func text() -> String! {
-        return self.messageText
-    }
-    
 }
