+++ ChatSecure/Classes/Utilities/OTRChatDemo.m
@@ -51,7 +51,7 @@
                 
             }
             
-            buddy.status = (NSInteger)OTRBuddyStatusAvailable+idx;
+            buddy.status = (NSInteger)ThreadStatusAvailable+idx;
             
             NSArray *textArray = [self shuffleHelloArray:helloArray];
