+++ ChatSecure/Classes/Utilities/OTRColors.m
@@ -10,24 +10,24 @@
 
 @implementation OTRColors
 
-+(UIColor *)colorWithStatus:(OTRBuddyStatus)status
++(UIColor *)colorWithStatus:(ThreadStatus)status
 {
     UIColor *color = nil;
     switch(status)
     {
-        case OTRBuddyStatusOffline:
+        case ThreadStatusOffline:
             color = [UIColor colorWithRed: 0.763 green: 0.763 blue: 0.763 alpha: 1];
             break;
-        case OTRBuddyStatusAway:
+        case ThreadStatusAway:
             color = [UIColor colorWithRed: 0.901 green: 0.527 blue: 0.23 alpha: 1];
             break;
-        case OTRBuddyStatusXa:
+        case ThreadStatusExtendedAway:
             color = [UIColor colorWithRed: 0.734 green: 0.124 blue: 0.124 alpha: 1];
             break;
-        case OTRBuddyStatusDnd:
+        case ThreadStatusDoNotDisturb:
             color = [UIColor colorWithRed: 0.734 green: 0.124 blue: 0.124 alpha: 1];
             break;
-        case OTRBuddyStatusAvailable:
+        case ThreadStatusAvailable:
             color = [UIColor colorWithRed: 0.083 green: 0.767 blue: 0.194 alpha: 1];
             break;
     }
